requirements.txt
llama-index
llama-index-agent-openai==0.2.7
llama-index-core==0.10.47
llama-index-embeddings-openai==0.1.10
llama-index-vector-stores-qdrant==0.2.10
qdrant-client==1.10.0
typesense==0.19.0
streamlit
python-dotenv
pandas
numpy